requirements.txt
# Fichier requirements.txt minimal pour les imports fournis
python-chess
torch
numpy
tqdm
requests
matplotlib